+++ luftdatenpumpe/core.py
@@ -10,10 +10,11 @@ import redis
 import logging
 
 from tqdm import tqdm
-from munch import Munch
+from munch import Munch, munchify
 from tablib import Dataset
 from requests_cache import CachedSession
 from luftdatenpumpe.geo import geohash_encode, resolve_location, improve_location, format_address
+from luftdatenpumpe.target import RDBMSStorage
 from luftdatenpumpe.util import exception_traceback, find_files
 
 from . import __appname__ as APP_NAME
@@ -70,6 +71,10 @@ class LuftdatenPumpe:
         return data
 
     def get_stations(self):
+
+        if self.source.startswith('postgresql://'):
+            return self.make_stations()
+
         stations = {}
         field_candidates = ['station_id', 'name', 'position', 'location']
         for reading in self.get_readings():
@@ -112,6 +117,96 @@ class LuftdatenPumpe:
             results.append(station)
 
         return results
+
+    def make_stations(self):
+        """
+        Re-make station information from PostgreSQL database.
+
+        Example::
+
+            {
+                "station_id": 28,
+                "name": "Ulmer Stra\u00dfe, Wangen, Stuttgart, Baden-W\u00fcrttemberg, DE",
+                "position": {
+                    "latitude": 48.778,
+                    "longitude": 9.236,
+                    "altitude": 223.7,
+                    "country": "DE",
+                    "geohash": "u0wt6pv2qqhz"
+                },
+                "location": {
+                    "place_id": "10504194",
+                    "licence": "Data \u00a9 OpenStreetMap contributors, ODbL 1.0. https://osm.org/copyright",
+                    "osm_type": "way",
+                    "osm_id": "115374184",
+                    "lat": "48.777845",
+                    "lon": "9.23582396156841",
+                    "display_name": "Kulturhaus ARENA, 241, Ulmer Stra\u00dfe, Wangen, Stuttgart, Regierungsbezirk Stuttgart, Baden-W\u00fcrttemberg, 70327, Deutschland",
+                    "boundingbox": [
+                        "48.7775199",
+                        "48.778185",
+                        "9.2353783",
+                        "9.236272"
+                    ],
+                    "address": {
+                        "country_code": "DE",
+                        "country": "Deutschland",
+                        "state": "Baden-W\u00fcrttemberg",
+                        "state_district": "Regierungsbezirk Stuttgart",
+                        "county": "Stuttgart",
+                        "postcode": "70327",
+                        "city": "Stuttgart",
+                        "city_district": "Wangen",
+                        "suburb": "Wangen",
+                        "road": "Ulmer Stra\u00dfe",
+                        "house_number": "241",
+                        "neighbourhood": "Wangen"
+                    },
+                    "address_more": {
+                        "building": "Kulturhaus ARENA"
+                    }
+                },
+                "sensors": [
+                    {
+                        "sensor_id": 658,
+                        "sensor_type": "SDS011"
+                    },
+                    {
+                        "sensor_id": 657,
+                        "sensor_type": "DHT22"
+                    }
+                ]
+            }
+        """
+
+        entries = []
+
+        sql = """
+            SELECT *
+            FROM ldi_stations, ldi_osmdata
+            WHERE
+              ldi_stations.id = ldi_osmdata.station_id
+            ORDER BY
+              ldi_stations.id
+        """
+
+        storage = RDBMSStorage(self.source)
+        for station in storage.db.query(sql):
+            station = munchify(station)
+            entry = {
+                "station_id": station.id,
+                "name": station.name,
+                "position": {
+                    "latitude": station.latitude,
+                    "longitude": station.longitude,
+                    "altitude": station.altitude,
+                    "country": station.country_code,
+                    "geohash": station.geohash
+                }
+            }
+            entries.append(munchify(entry))
+
+        return entries
 
     def request_live_data(self):